IGNITE-16964 SQL API: Implement async SQL API (#817)

+++ modules/api/src/main/java/org/apache/ignite/sql/reactive/ReactiveResultSet.java
@@ -58,7 +58,7 @@ public interface ReactiveResultSet extends Flow.Publisher<SqlRow> {
      * @return Publisher for number of rows.
      * @see ResultSet#affectedRows()
      */
-    Flow.Publisher<Integer> affectedRows();
+    Flow.Publisher<Long> affectedRows();
 
     /**
      * Returns publisher for a flag which determines whether the query that produce this result was a conditional query, or not.
